Termino Observers de Notificacion, cambio strategy Interesada para que reciba valorTope

+++ src/main/kotlin/PreferenciaPorRegalos.kt
@@ -10,9 +10,10 @@ class PersonaConformista : PreferenciaPorRegalos {
 }
 
 //Quieren que el regalo cueste más de una cierta cantidad de $$$
-class PersonaInteresada(val montoTope : Int) : PreferenciaPorRegalos {
+//Nunca se habia mencionado un MONTO TOPE!!!!!
+class PersonaInteresada(val montoTope : Double) : PreferenciaPorRegalos {
     override fun aceptaRegalo(regalo: Regalo): Boolean {
-        val precioBase = 5000
+        val precioBase = 5000.0
         return regalo.valor >= precioBase && regalo.valor <= montoTope  //Modificado en base a Punto 3 de la consigna: define una base dentro y luego un tope por constructor.
     }
 }
